refactor(*): camelCase
