+++ app/src/main/java/com/kristallik/jokeapp/presentation/ui/jokeDetails/fragment/JokeDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.kristallik.jokeapp.presentation.ui.joke_details.fragment
+package com.kristallik.jokeapp.presentation.ui.jokeDetails.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,8 +8,8 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import com.kristallik.jokeapp.domain.model.Joke
 import com.kristallik.jokeapp.databinding.FragmentJokeDetailsBinding
-import com.kristallik.jokeapp.presentation.ui.joke_details.JokeDetailsPresenter
-import com.kristallik.jokeapp.presentation.ui.joke_details.JokeDetailsView
+import com.kristallik.jokeapp.presentation.ui.jokeDetails.JokeDetailsPresenter
+import com.kristallik.jokeapp.presentation.ui.jokeDetails.JokeDetailsView
 
 class JokeDetailsFragment : Fragment(), JokeDetailsView {
     private var _binding: FragmentJokeDetailsBinding? = null
